requirements.txt
Django==5.1.3
djangorestframework==3.15.2
drf_yasg==1.21.8
python-dotenv==1.0.1
